# 8261744: Implement CharsetDecoder ASCII and latin-1 fast-paths