--- a/src/java.base/share/classes/java/lang/System.java
+++ b/src/java.base/share/classes/java/lang/System.java
@@ -2005,6 +2005,10 @@ public final class System {
 
         lineSeparator = props.getProperty("line.separator");
 
+        // register shared secrets - do this before setting up System.out/in
+        // since encoders/decoders rely on it
+        setJavaLangAccess();
+
         FileInputStream fdIn = new FileInputStream(FileDescriptor.in);
         FileOutputStream fdOut = new FileOutputStream(FileDescriptor.out);
         FileOutputStream fdErr = new FileOutputStream(FileDescriptor.err);
@@ -2026,8 +2030,6 @@ public final class System {
         Thread current = Thread.currentThread();
         current.getThreadGroup().add(current);
 
-        // register shared secrets
-        setJavaLangAccess();
 
         // Subsystems that are invoked during initialization can invoke
         // VM.isBooted() in order to avoid doing things that should