--- a/src/java.base/share/classes/sun/nio/cs/CESU_8.java
+++ b/src/java.base/share/classes/sun/nio/cs/CESU_8.java
@@ -25,6 +25,9 @@
 
 package sun.nio.cs;
 
+import jdk.internal.access.JavaLangAccess;
+import jdk.internal.access.SharedSecrets;
+
 import java.nio.Buffer;
 import java.nio.ByteBuffer;
 import java.nio.CharBuffer;
@@ -75,6 +78,9 @@ class CESU_8 extends Unicode
 
     private static class Decoder extends CharsetDecoder
                                  implements ArrayDecoder {
+
+        private static final JavaLangAccess JLA = SharedSecrets.getJavaLangAccess();
+
         private Decoder(Charset cs) {
             super(cs, 1.0f, 1.0f);
         }
@@ -94,27 +100,6 @@ class CESU_8 extends Unicode
         private static boolean isMalformed3_2(int b1, int b2) {
             return (b1 == (byte)0xe0 && (b2 & 0xe0) == 0x80) ||
                    (b2 & 0xc0) != 0x80;
-        }
-
-
-        //  [F0]     [90..BF] [80..BF] [80..BF]
-        //  [F1..F3] [80..BF] [80..BF] [80..BF]
-        //  [F4]     [80..8F] [80..BF] [80..BF]
-        //  only check 80-be range here, the [0xf0,0x80...] and [0xf4,0x90-...]
-        //  will be checked by Character.isSupplementaryCodePoint(uc)
-        private static boolean isMalformed4(int b2, int b3, int b4) {
-            return (b2 & 0xc0) != 0x80 || (b3 & 0xc0) != 0x80 ||
-                   (b4 & 0xc0) != 0x80;
-        }
-
-        // only used when there is less than 4 bytes left in src buffer
-        private static boolean isMalformed4_2(int b1, int b2) {
-            return (b1 == 0xf0 && b2 == 0x90) ||
-                   (b2 & 0xc0) != 0x80;
-        }
-
-        private static boolean isMalformed4_3(int b3) {
-            return (b3 & 0xc0) != 0x80;
         }
 
         private static CoderResult malformedN(ByteBuffer src, int nb) {
@@ -202,17 +187,19 @@ class CESU_8 extends Unicode
         {
             // This method is optimized for ASCII input.
             byte[] sa = src.array();
-            int sp = src.arrayOffset() + src.position();
-            int sl = src.arrayOffset() + src.limit();
+            int soff = src.arrayOffset();
+            int sp = soff + src.position();
+            int sl = soff + src.limit();
 
             char[] da = dst.array();
-            int dp = dst.arrayOffset() + dst.position();
-            int dl = dst.arrayOffset() + dst.limit();
-            int dlASCII = dp + Math.min(sl - sp, dl - dp);
+            int doff = dst.arrayOffset();
+            int dp = doff + dst.position();
+            int dl = doff + dst.limit();
 
-            // ASCII only loop
-            while (dp < dlASCII && sa[sp] >= 0)
-                da[dp++] = (char) sa[sp++];
+            int n = JLA.decodeASCII(sa, sp, da, dp, Math.min(sl - sp, dl - dp));
+            sp += n;
+            dp += n;
+
             while (sp < sl) {
                 int b1 = sa[sp];
                 if (b1 >= 0) {