… src/java.base/share/classes/sun/nio/cs/SingleByte.java
… src/java.base/share/classes/sun/nio/cs/StreamDecoder.java
… src/java.base/share/classes/sun/nio/cs/StreamEncoder.java
… src/java.base/share/classes/sun/nio/cs/US_ASCII.java
… src/java.base/share/classes/sun/nio/cs/UTF_8.java
… test/micro/org/openjdk/bench/java/io/ByteStreamDecoder.java